+++ examples/SimultaneousButtons/SimultaneousButtons.ino
@@ -154,8 +154,10 @@ ButtonHandler handleEvent(true /* requireConcurrentTrigger*/);
 void checkButtons() {
   static uint16_t lastCheck;
 
+  // DO NOT USE delay(5) to do this.
+  // The (uint16_t) cast is required on 32-bit processors, harmless on 8-bit.
   uint16_t now = millis();
-  if ((now - lastCheck) >= 5) {
+  if ((uint16_t) (now - lastCheck) >= 5) {
     lastCheck = now;
     b1.check();
     b2.check();
